+++ lib/screens/ocr_result_item_widgets/ocr_result_empty_state.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+
+/// OCR結果アイテムが0件の場合の空状態ウィジェット
+class OcrResultEmptyState extends StatelessWidget {
+  const OcrResultEmptyState({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+
+    return Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Icon(
+            Icons.inbox_outlined,
+            size: 64,
+            color: theme.colorScheme.outline,
+          ),
+          const SizedBox(height: 16),
+          Text(
+            '商品がありません',
+            style: theme.textTheme.titleMedium?.copyWith(
+              color: theme.colorScheme.outline,
+            ),
+          ),
+          const SizedBox(height: 8),
+          Text(
+            '撮影した画像から商品を検出できませんでした',
+            style: theme.textTheme.bodyMedium?.copyWith(
+              color: theme.colorScheme.outline,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
