+++ lib/screens/ocr_result_item_widgets/ocr_result_save_button.dart
@@ -1,0 +1,59 @@
+import 'package:flutter/material.dart';
+
+/// OCR結果の保存ボタンウィジェット
+class OcrResultSaveButton extends StatelessWidget {
+  const OcrResultSaveButton({
+    super.key,
+    required this.isProcessing,
+    required this.onSave,
+  });
+
+  final bool isProcessing;
+  final VoidCallback onSave;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+
+    return Container(
+      padding: EdgeInsets.fromLTRB(
+        16,
+        8,
+        16,
+        8 + MediaQuery.of(context).padding.bottom,
+      ),
+      child: SizedBox(
+        width: double.infinity,
+        height: 48,
+        child: ElevatedButton(
+          onPressed: isProcessing ? null : onSave,
+          style: ElevatedButton.styleFrom(
+            backgroundColor: theme.colorScheme.primary,
+            foregroundColor: theme.colorScheme.onPrimary,
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(12),
+            ),
+          ),
+          child: isProcessing
+              ? SizedBox(
+                  width: 24,
+                  height: 24,
+                  child: CircularProgressIndicator(
+                    strokeWidth: 2,
+                    valueColor: AlwaysStoppedAnimation<Color>(
+                      theme.colorScheme.onPrimary,
+                    ),
+                  ),
+                )
+              : Text(
+                  '保存する',
+                  style: TextStyle(
+                    fontSize: theme.textTheme.bodyLarge?.fontSize,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+        ),
+      ),
+    );
+  }
+}
